+++ videoplayer-api/src/main/java/intbird/soft/lib/video/player/api/error/MediaError.kt
@@ -1,0 +1,19 @@
+package intbird.soft.lib.video.player.api.error
+
+/**
+ * created by Bird
+ * on 2020/9/8
+ * DingTalk id: intbird
+ */
+object MediaError {
+    const val NO_ERROR = 0
+
+    const val NO_MORE_FILES = 1
+
+    const val PLAYER_INIT_ERROR = 10
+    const val PLAYER_PREPARE_ERROR = 11
+    const val PLAYER_ERROR_CALLBACK = 12
+
+
+}
+